pipeline-troubleshooting-assistant/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
pydantic==1.10.12
pyyaml==6.0
pytest==7.4.0
requests==2.31.0